// src/pages/components/queries/queries.generated.ts
export type HeroImageFragment = { id: string, altText?: string | null, fileName: string, top3000?: string | null, top2400?: string | null, top1600?: string | null, top1400?: string | null, top1200?: string | null, top1140?: string | null, top1024?: string | null, top800?: string | null, top768?: string | null, top500?: string | null, top480?: string | null };